Added framework for issue regarding copy and paste
This copy and paste issue has been going on for a while. When copying and pasting while the code is folded, upon pasting it unfolds automatically. I'm not that familiar with your code so I will implement it later once I have time.

Framework for copy and paste issue on brackets main github
        https://github.com/adobe/brackets/
        Issue #13014